requirements.txt
# requirements.txt
functions-framework
numpy
scikit-learn
torch
transformers
google-cloud-storage